Use alternate function definition style for init-path
Starting in zsh 5.4, something in the chain of how init-path is defined
broke, so that it only worked as expected on its second invocation.
This change removes the "init-path () { ... }" function definition
syntax, instead putting the bare function body as the top-level contents
of the file.  This appears to satisfy whatever breaking change occurred
in zsh 5.4.